feat: Netex for circular/non-circular returns
