+++ src/netex-convertor/handler.ts
@@ -1,6 +1,6 @@
 import { S3Event } from 'aws-lambda';
-import singleTicketNetexGenerator from './single-ticket/singleTicketNetexGenerator';
-import periodTicketNetexGenerator from './period-ticket/periodTicketNetexGenerator';
+import pointToPointTicketNetexGenerator from './point-to-point-tickets/pointToPointTicketNetexGenerator';
+import periodTicketNetexGenerator from './period-tickets/periodTicketNetexGenerator';
 import * as db from './data/auroradb';
 import * as s3 from './data/s3';
 import { MatchingData, PeriodTicket } from './types';
@@ -13,11 +13,11 @@ export const netexConvertorHandler = async (event: S3Event): Promise<void> => {
 
         console.info(`NeTEx generation starting for type: ${type}...`);
 
-        if (type === 'pointToPoint') {
+        if (type === 'pointToPoint' || type === 'return') {
             const matchingData: MatchingData = s3Data;
             const operatorData = await db.getOperatorDataByNocCode(matchingData.nocCode);
 
-            const netexGen = singleTicketNetexGenerator(matchingData, operatorData);
+            const netexGen = pointToPointTicketNetexGenerator(matchingData, operatorData);
             const generatedNetex = await netexGen.generate();
 
             const fileName = `${matchingData.operatorShortName.replace(/\/|\s/g, '_')}_${
