+++ src/netex-convertor/period-tickets/periodTicketNetexTemplate.xml
