+++ src/netex-convertor/test-data/geoZonePeriodData.ts
