+++ src/netex-convertor/test-data/operatorData.ts
